@@ -99,12 +100,12 @@
     "pdfjs-dist": "^1.10.97",
     "points-cluster": "^0.1.4",
     "postcss-browser-reporter": "^0.5.0",
-    "postcss-cssnext": "^2.10.0",
-    "postcss-loader": "^1.3.3",
-    "postcss-reporter": "^3.0.0",
-    "promise-polyfill": "^6.1.0",
-    "prop-types": "^15.5.10",
-    "react": "^15.6.1",
+    "postcss-cssnext": "^3.0.2",
+    "postcss-loader": "^2.0.9",
+    "postcss-reporter": "^5.0.0",
+    "promise-polyfill": "^7.0.0",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
     "react-addons-css-transition-group": "^15.6.2",
     "react-addons-shallow-compare": "^15.6.2",
     "react-avatar": "^2.5.1",
@@ -115,86 +116,83 @@
     "react-day-picker": "^5.5.3",
     "react-dnd": "^2.5.4",
     "react-dnd-html5-backend": "^2.5.4",
-    "react-dom": "^15.4.2",
-    "react-dropzone": "github:izadmehr/react-dropzone#eb84dc41f26c6d02698b321fe3c42f3114495674",
-    "react-ga": "^2.2.0",
+    "react-dom": "^16.2.0",
+    "react-dropzone":
+      "github:izadmehr/react-dropzone#eb84dc41f26c6d02698b321fe3c42f3114495674",
+    "react-ga": "^2.3.5",
     "react-hot-loader": "^3.1.3",
     "react-images": "^0.5.11",
-    "react-input-mask": "^1.0.7",
-    "react-intercom": "^1.0.13",
-    "react-ios-switch": "0.1.10",
-    "react-lazy-load": "^3.0.12",
+    "react-input-mask": "^1.1.0",
+    "react-intercom": "^1.0.14",
+    "react-lazy-load": "^3.0.13",
     "react-loadable": "^5.3.1",
-    "react-maskedinput": "^3.3.4",
-    "react-motion": "^0.5.0",
-    "react-redux": "^5.0.4",
+    "react-maskedinput": "^4.0.0",
+    "react-motion": "^0.5.2",
+    "react-redux": "^5.0.6",
     "react-resize-detector": "^0.6.0",
     "react-router": "^3.0.2",
     "react-router-bootstrap": "^0.23.1",
     "react-router-redux": "^4.0.8",
-    "react-select": "1.0.0-rc.5",
-    "react-swipe": "^5.0.6",
+    "react-select": "1.1.0",
+    "react-swipe": "5.1.1",
     "react-telephone-input": "^4.3.4",
     "react-text-mask": "^5.0.1",
     "react-textarea-autosize": "^5.2.1",
     "read-chunk": "^2.0.0",
     "reapop": "^1.1.4",
     "reapop-theme-wybo": "^0.4.1",
-    "recompose": "^0.23.1",
-    "redux": "^3.6.0",
-    "redux-auth-wrapper": "^2.0.1",
+    "recompose": "^0.26.0",
+    "redux": "^3.7.2",
+    "redux-auth-wrapper": "^2.0.2",
     "redux-batched-actions": "^0.2.0",
     "redux-devtools-extension": "^2.13.2",
     "redux-form": "^7.2.0",
     "redux-thunk": "^2.2.0",
     "rimraf": "^2.6.1",
-    "rxjs": "^5.5.2",
-    "sass-loader": "^6.0.3",
+    "rxjs": "^5.5.6",
+    "sass-loader": "^6.0.6",
     "scissors": "git+https://github.com/tcr/scissors.git",
     "shorti": "^1.2.6",
     "socket.io": "^1.7.3",
     "socket.io-client": "^1.7.3",
     "stream-to-promise": "^2.2.0",
-    "style-loader": "^0.16.1",
-    "superagent": "^3.8.1",
+    "style-loader": "^0.19.1",
+    "superagent": "^3.8.2",
     "swipe-js-iso": "^2.0.3",
     "text-mask-addons": "^3.7.1",
     "underscore": "^1.8.3",
     "universal-cookie": "^2.1.2",
-    "uppy": "^0.15.0",
-    "uuid-validate": "0.0.2",
-    "validator": "^7.0.0",
-    "webpack": "^3.8.1",
+    "validator": "^9.2.0",
+    "webpack": "^3.10.0",
     "webpack-bundle-analyzer": "^2.9.1",
     "webpack-visualizer-plugin": "^0.1.11",
     "whatwg-fetch": "^2.0.3",
-    "xss": "^0.3.3"
+    "xss": "^0.3.6"
   },
   "devDependencies": {
-    "babel-eslint": "^7.2.3",
-    "babel-jest": "^21.2.0",
+    "babel-eslint": "^8.1.0",
+    "babel-jest": "^22.0.4",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-preset-stage-2": "^6.24.1",
     "enzyme": "^3.1.1",
-    "eslint": "^4.12.0",
-    "eslint-config-airbnb": "^15.1.0",
+    "eslint": "^4.14.0",
+    "eslint-config-airbnb": "^16.1.0",
     "eslint-loader": "^1.9.0",
     "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.5.1",
     "fetch-mock": "^5.13.1",
-    "jest": "^21.2.1",
-    "jsdom": "^9.12.0",
+    "jest": "^22.0.4",
+    "jsdom": "^11.5.1",
     "nightwatch": "^0.9.16",
     "nodemon": "^1.11.0",
-    "react-test-renderer": "^16.1.1",
+    "react-test-renderer": "^16.2.0",
     "redux-mock-store": "^1.3.0",
     "selenium-standalone": "^6.11.0",
-    "should": "^11.2.1",
     "uglify-es": "git://github.com/mishoo/UglifyJS2.git#harmony",
-    "uglifyjs-webpack-plugin": "^1.1.1",
-    "webpack-dev-middleware": "^1.12.2",
-    "webpack-dev-server": "^2.9.5",
+    "uglifyjs-webpack-plugin": "^1.1.4",
+    "webpack-dev-middleware": "^2.0.3",
+    "webpack-dev-server": "^2.9.7",
     "webpack-hot-middleware": "^2.21.0",
     "webpack-notifier": "^1.5.0"
   }
