@@ -1,13 +1,9 @@
 import React from 'react'
-import pure from 'recompose/pure'
 
-const Accordion = ({ active, children }) =>
-  <div
-    className={`c-filters-status__accordion-target ${active
-      ? 'is-active'
-      : ''}`}
-  >
+const Accordion = ({ active, children }) => (
+  <div className={`c-filters-status__accordion-target ${active ? 'is-active' : ''}`}>
     {children}
   </div>
+)
 
-export default pure(Accordion)
+export default Accordion
