@@ -3,31 +3,26 @@ import compose from 'recompose/compose'
 import defaultProps from 'recompose/defaultProps'
 // import withHandlers from 'recompose/withHandlers'
 import withPropsOnChange from 'recompose/withPropsOnChange'
-import pure from 'recompose/pure'
 import { Motion, spring } from 'react-motion'
 
 export const clusterMarker = ({
   text,
   onClickHandler,
-  defaultMotionStyle, motionStyle
+  defaultMotionStyle,
+  motionStyle
 }) => (
-  <Motion
-    style={motionStyle}
-    defaultStyle={defaultMotionStyle}
-  >
-    {
-      ({ scale }) => (
-        <div
-          className="cluster-marker"
-          onClick={onClickHandler}
-          style={{
-            transform: `translate3D(0,0,0) scale(${scale}, ${scale})`
-          }}
-        >
-          <div>{text}</div>
-        </div>
-      )
-    }
+  <Motion style={motionStyle} defaultStyle={defaultMotionStyle}>
+    {({ scale }) => (
+      <div
+        className="cluster-marker"
+        onClick={onClickHandler}
+        style={{
+          transform: `translate3D(0,0,0) scale(${scale}, ${scale})`
+        }}
+      >
+        <div>{text}</div>
+      </div>
+    )}
   </Motion>
 )
 
@@ -42,9 +37,6 @@ export const clusterMarkerHOC = compose(
     damping: 7,
     precision: 0.001
   }),
-  // pure optimization can cause some effects you don't want,
-  // don't use it in development for markers
-  pure,
   withPropsOnChange(
     ['initialScale'],
     ({ initialScale, defaultScale, $prerender }) => ({
@@ -54,16 +46,14 @@ export const clusterMarkerHOC = compose(
   ),
   withPropsOnChange(
     ['hovered'],
-    ({
-      hovered, hoveredScale, defaultScale,
-      stiffness, damping, precision
-    }) => ({
+    ({ hovered, hoveredScale, defaultScale, stiffness, damping, precision }) => ({
       hovered,
       motionStyle: {
-        scale: spring(
-          hovered ? hoveredScale : defaultScale,
-          { stiffness, damping, precision }
-        )
+        scale: spring(hovered ? hoveredScale : defaultScale, {
+          stiffness,
+          damping,
+          precision
+        })
       }
     })
   )
