@@ -1,5 +1,5 @@
 import Koa from 'koa'
-import validate from 'uuid-validate'
+
 const router = require('koa-router')()
 const app = new Koa()
 
@@ -7,12 +7,13 @@ router.get('/listings/:id', async (ctx, next) => {
   const { id } = ctx.params
 
   // Catch non-uuid
-  if (!validate(id)) { return await next() }
+  if (!validate(id)) {
+    return await next()
+  }
 
   try {
-    const response = await ctx.fetch(
-      `/listings/${id}?associations[]=listing.proposed_agent`
-    )
+    const response = await ctx.fetch(`/listings/${id}?associations[]=listing.proposed_agent`)
+
     ctx.body = response.body
   } catch (e) {
     console.log(e)
